Introduce gnar styling preferences
This adds in two YAML files that may be inherited from when adhering to
Gnar's style preferences. The first, default, is the base level for all
ruby projects. default-rails adds in checks that are specific for a
rails project.

This also includes a .rubocop.yml that inherits from default.yml both as
an example of how to use these defaults and also to ensure this repo
itself adheres to the style conventions of The Gnar Company.