Separate Trailing Comma Array and Hash Literal
Rubocop 0.53.0 created two new cops for controlling trailing comma style
preferences for arrays and hashes which may be found here:
https://github.com/bbatsov/rubocop/pull/5307.

This removes the `Style/TrailingCommaInLiteral` cop in favor of these
more granular cops.

As a result of this change, gnar-style needs to be updated to use the
new cops and remove the use of the old one. This updates to use the new
cops and updates the gemspec to reflect that this now requires 0.53 of
rubocop in this version to adhere to the available cops.