Rename IndentArray cop
In rubocop [0.68.0](https://github.com/rubocop-hq/rubocop/blob/master/CHANGELOG.md#0680-2019-04-29),
the `IndentArray` cop was renamed to `IndentFirstArrayElement`. This
change supports that name change by modifying our style principle for
this cop to be named appropriately.

This also updates the rubocop version that this is dependent on, as that
name change first appears in 0.68.0, so for rubocop to understand what
to do with that cop, we must have that version.