check in Gemfile.lock
Bundler
[recommends](https://bundler.io/blog/2018/01/17/making-gem-development-a-little-better.html)
checking in `Gemfile.lock` even when developing a gem. It makes it much
easier for outside developers to contribute.

This is doubly true with Bundler 2.0. Bundler 2.0
[checks](https://bundler.io/guides/bundler_2_upgrade.html) the `BUNDLED
WITH` part of the `Gemfile.lock` to decide which version of Bundler to
run. Without `Gemfile.lock`, running `bundle install`
produces the error:

`Bundler could not find compatible versions for gem "bundler":`

This means that if `Gemfile.lock` is not checked in, a first-time contributor
has to install Bundler 1.14, uninstall all other versions of Bundler, run
`bundle install` and then reinstall the newer version of Bundler (which will
now work because they have a local version of `Gemfile.lock`).